Simplified participant API
Let's try and keep the participant object immutable. At this point there
is no need to mutate a participant's player character and so we can just
treat it as its ID.

Of course because of this there will be problems further down the road
(e.g.: What to do about duplicate player IDs?) but we will understand
those problems better when we're there.