Refactoring and minor improvement to API
We are collecting data structures and functions that will be used by
Participants as well as Referees in the actors package. We also turned
spies into private types. There's no need to make them available
anywhere else. And we created stubs for possible responses to our
requests to the Referee.