Tell, Don't Ask: First step in reducing complexity
Most of the current complexity is due to us constantly asking for this
or that value. Even though x.Validate() and its helper methods aren't
bad per se - they are actually necessary for want of dependable types -
they should be used sparsely. Instead of checking if a given Game state
is legal we will prioritize eliminating functionality that might turn it
illegal.