Tell, don't ask: Game should be a mutable object after all
We can't have Referee and Participant all constantly being occupied with
figuring out the current state of the game. Instead we'll have a single
way to create a Game, New(Size) (*Game, error), and a limited set of
methods to change its state safely. This way we can also toss out all
the validation methods and let Referee be concerned with managing
Participants' access to the Game and keeping them Up-To-Date.